+++ accountant/accountant-core/src/main/kotlin/co/nilin/opex/accountant/core/spi/PairStaticRateLoader.kt
@@ -1,5 +1,7 @@
 package co.nilin.opex.accountant.core.spi
 
+import java.math.BigDecimal
+
 interface PairStaticRateLoader {
-    suspend fun calculateStaticRate(leftSide: String, rightSide: String): Double?
+    suspend fun calculateStaticRate(leftSide: String, rightSide: String): BigDecimal?
 }
