+++ accountant/accountant-ports/accountant-persister-postgres/src/main/kotlin/co/nilin/opex/accountant/ports/postgres/model/PairConfigModel.kt
@@ -3,12 +3,13 @@ package co.nilin.opex.accountant.ports.postgres.model
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Column
 import org.springframework.data.relational.core.mapping.Table
+import java.math.BigDecimal
 
 @Table("pair_config")
 data class PairConfigModel(
     @Id val pair: String,
     @Column("left_side_wallet_symbol") val leftSideWalletSymbol: String, //can be same as pair left side
     @Column("right_side_wallet_symbol") val rightSideWalletSymbol: String, //can be same as pair right side
-    @Column("left_side_fraction") val leftSideFraction: Double,
-    @Column("right_side_fraction") val rightSideFraction: Double
+    @Column("left_side_fraction") val leftSideFraction: BigDecimal,
+    @Column("right_side_fraction") val rightSideFraction: BigDecimal
 )
