+++ api/api-core/src/main/kotlin/co/nilin/opex/api/core/inout/PairFeeResponse.kt
@@ -1,9 +1,11 @@
 package co.nilin.opex.api.core.inout
 
+import java.math.BigDecimal
+
 data class PairFeeResponse(
-    val pair:String,
+    val pair: String,
     val direction: String,
     val userLevel: String,
-    val makerFee: Double,
-    val takerFee: Double
+    val makerFee: BigDecimal,
+    val takerFee: BigDecimal
 )
