+++ api/api-core/src/main/kotlin/co/nilin/opex/api/core/inout/PairInfoResponse.kt
@@ -1,9 +1,11 @@
 package co.nilin.opex.api.core.inout
 
+import java.math.BigDecimal
+
 data class PairInfoResponse(
     val pair: String,
     val leftSideWalletSymbol: String,
     val rightSideWalletSymbol: String,
-    val leftSideFraction: Double,
-    val rightSideFraction: Double
+    val leftSideFraction: BigDecimal,
+    val rightSideFraction: BigDecimal
 )
