+++ websocket/websocket-core/src/main/kotlin/co/nilin/opex/websocket/core/inout/AggregatedOrderPriceModel.kt
@@ -1,6 +1,8 @@
 package co.nilin.opex.websocket.core.inout
 
+import java.math.BigDecimal
+
 data class AggregatedOrderPriceModel(
-    val price: Double?,
-    val quantity: Double?
+    val price: BigDecimal?,
+    val quantity: BigDecimal?
 )
