+++ websocket/websocket-ports/websocket-persister-postgres/src/main/kotlin/co/nilin/opex/websocket/ports/postgres/model/OrderModel.kt
@@ -8,6 +8,7 @@ import org.springframework.data.annotation.Id
 import org.springframework.data.annotation.Version
 import org.springframework.data.relational.core.mapping.Column
 import org.springframework.data.relational.core.mapping.Table
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 @Table("orders")
@@ -20,17 +21,17 @@ class OrderModel(
     val clientOrderId: String?,
     val symbol: String,
     @Column(value = "order_id") val orderId: Long?,
-    @Column("maker_fee") val makerFee: Double?,
-    @Column("taker_fee") val takerFee: Double?,
-    @Column("left_side_fraction") val leftSideFraction: Double?,
-    @Column("right_side_fraction") val rightSideFraction: Double?,
+    @Column("maker_fee") val makerFee: BigDecimal?,
+    @Column("taker_fee") val takerFee: BigDecimal?,
+    @Column("left_side_fraction") val leftSideFraction: BigDecimal?,
+    @Column("right_side_fraction") val rightSideFraction: BigDecimal?,
     @Column("user_level") val userLevel: String?,
     @Column("side") val direction: OrderDirection?,
     @Column("match_constraint") val constraint: MatchConstraint?,
     @Column("order_type") val type: OrderType?,
-    @Column("price") val price: Double?,
-    @Column("quantity") val quantity: Double?,
-    @Column("quote_quantity") val quoteQuantity: Double?,
+    @Column("price") val price: BigDecimal?,
+    @Column("quantity") val quantity: BigDecimal?,
+    @Column("quote_quantity") val quoteQuantity: BigDecimal?,
     @Column("create_date") val createDate: LocalDateTime?,
     @Column("update_date") val updateDate: LocalDateTime,
     @Version
